[unpack] Add unpacking logic for git based source
