+++ internal/controllers/v1alpha2/source/image.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"compress/gzip"
 	"context"
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -294,4 +295,34 @@ func pendingImagePodResult(pod *corev1.Pod) *Result {
 		}
 	}
 	return &Result{State: StatePending, Message: strings.Join(messages, "; ")}
+}
+
+// getCachedContentPath returns the name of the cached directory if exists.
+func getCachedContentPath(bdaName string, bundleSrc *v1alpha2.BundleDeplopymentSource, base afero.Fs) (string, error) {
+	cachedDirName := getCacheDirName(bdaName, *bundleSrc)
+
+	if ok, err := afero.DirExists(base, filepath.Join(CacheDir, cachedDirName)); err != nil {
+		return "", fmt.Errorf("error finding cache dir %v", err)
+	} else if !ok {
+		return "", nil
+	}
+	return cachedDirName, nil
+}
+
+// Perform a base64 encoding to get the directoryName to store caches
+func getCacheDirName(bdName string, bd v1alpha2.BundleDeplopymentSource) string {
+	switch bd.Kind {
+	case v1alpha2.SourceTypeImage:
+		return encode(bdName, string(bd.Kind), bd.Image.ImageRef)
+	case v1alpha2.SourceTypeGit:
+		return encode(bdName, string(bd.Kind), bd.Git.Repository)
+	case v1alpha2.SourceTypeHTTP:
+		return encode(bdName, string(bd.Kind), bd.HTTP.URL)
+	default:
+		return ""
+	}
+}
+
+func encode(str1, str2, str3 string) string {
+	return base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s-%s-%s", str1, str2, str3)))
 }
